+++ src/dlc/platforms/AmagProTime/AmagProTime.js
@@ -6,7 +6,6 @@ import {
   filterBookingNomber
 } from "./services/AmagProTime.services.js";
 import { message } from "../../../components/ui/message/message.js";
-import { notification } from "../../../components/ui/notification/notification.js";
 import {
   bookingLoopCount,
   highLatency,